Project/requirements.txt
Pillow>=9.0.0
requests>=2.25.1
pytest>=7.0.0